+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Change log
 
-## Devel
+## 0.9 (2023-01-09)
 
 - Bug fix: Do not allow `ManagedGroup` or `Workspace` instances to be deleted if referenced in another model by a protected foreign key.
 - Add a customizable adapter for `Accounts`. This requires setting the `ANVIL_ACCOUNT_ADAPTER` setting in your settings file.
-- Various UI updates: the detail box on object detail pages now uses a description list instead of an onordered list.
+- Various UI updates: the detail box on object detail pages now uses a description list instead of an unordered list.
 
 ## 0.8 (2022-12-27)